requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
python-decouple>=3.8
pydantic>=2.7.0,<3.0.0
agents>=0.1.0
